requirements.txt
# Core FastAPI
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.7.4
python-multipart==0.0.6

# Database
supabase==2.3.0

# LLM Integration
langchain-groq==0.0.1
langchain-core==0.1.53
langgraph==0.0.26
langsmith==0.1.147

# Environment Variables
python-dotenv==1.0.0

# PDF Generation
reportlab==4.0.7

# Additional utilities
requests==2.31.0
beautifulsoup4==4.12.2
